travis: Make expected result an `env` variable
In an attempt to improve isolation of tests from the testing machine in
Travis CI and improve support for a wider range of Linux distributions,
a couple tests will be added which run in containers. Some of them will
fail at the beginning because we will use a number of Linux
distributions' userlands which aren't supported yet to 100%. To ensure
the test jobs are working well, they need to fail before fixes are
implemented (red-green testing).

To accommodate for red-green testing, add an `expected_result`
environment variable which will be set per job later.

It's a bit tricky because there are two cases, _equals zero_ and _does
not equal zero_. To encode the operator and keep build matrix readable,
introduce the `expected_result` environment variable which is a quoted
string, inserted in an `eval` statement later.

Add `_job` mapping as the base for `_native_job`, so later, when
`_docker_job` will be added, we don't need to repeat ourselves.